requirements.txt
Flask
opencv-python
Pillow
numpy